requirements.txt
# Python 3.8

certifi==2022.6.15
cffi==1.14.2
charset-normalizer==2.0.12
click==8.1.3
docker-pycreds==0.4.0
gitdb==4.0.9
GitPython==3.1.27
idna==3.3
numpy==1.22.4
pandas==1.4.2
pathtools==0.1.2
Pillow==9.1.1
promise==2.3
protobuf==3.20.1
psutil==5.9.1
pycparser==2.21
PyRep==4.1.0.3
python-dateutil==2.8.2
pytz==2022.1
PyYAML==6.0
requests==2.28.0
ruamel.yaml==0.17.21
ruamel.yaml.clib==0.2.6
sentry-sdk==1.9.2
setproctitle==1.3.1
shortuuid==1.0.9
six==1.16.0
smmap==5.0.0
torch==1.11.0+cu113
torchaudio==0.11.0+cu113
torchvision==0.12.0+cu113
typing-extensions==4.2.0
urllib3==1.26.9
wandb==0.13.1
